Phase K Wave 67: case-8 sizeof(long) preservation-fidelity tightening + byte-exact 4B 0x0096 hardening test
case-8 of Player::HandleStarbaseRequest at PlayerConnection.cpp:10006-10020
used to emit 0x0096 JOB_ACCEPT_REPLY with sizeof(long) bytes. The retail
Win32 server emitted 4 bytes (LP32 Win32 long is 4-byte). Our Linux x86_64
server was emitting 8 bytes (LP64 Linux long is 8-byte), drifting the wire
shape away from the retail format and risking inner-opcode mis-frame inside
the 0x2016 PACKET_SEQUENCE the case-8 reply rides in.

Fix: a single token swap — `long job_id` -> `int32_t job_id` at
PlayerConnection.cpp:10012 — locks sizeof(job_id) to 4 on both platforms.

Per CLAUDE.md server-integrity rule "Server integrity rules", this is
exactly the kind of change the rule welcomes: "It does NOT restrict changes
that *tighten* the server toward greater fidelity ... those are always
welcome." Concretely:

  - No widened input acceptance (we still process 9-byte STARBASE_REQUEST
    payloads identically; only the OUTBOUND wire shape changed).
  - No loosened gating (case-8 still requires the JT venue + Action=8).
  - No fabricated replies (the emit site still exists in the upstream
    server unchanged; only the local var's storage class changed).
  - No relaxed posture (AwardJob still fires and the trailing
    SendVaMessageC unavailable-message fallback still runs unchanged).

Hardening test SectorStarbaseAcceptJob8Tests.StarbaseAcceptJobAction8_OnSentinelJobId_ReceivesByteExact4ByteJobAcceptReply
pins the tightening:

  - Two-stage handshake EstablishAsync(10151) + ReestablishAsync(10711)
    into Net-7 SOL (the unique JT-equipped starbase in the seed).
  - Stimulus: 9B 0x004E STARBASE_REQUEST with Action=8 (Accept Job) and
    StarbaseID=0x12345678 (sentinel job_id provably absent from m_JobList
    since RefreshJobs assigns IDs monotonically from 0 via m_JobListID++).
  - Asserts payload.Length == 4 EXACTLY (load-bearing — pins the tightening)
    and int32 LE at [0..4] == 0x12345678 (sentinel echo verbatim).

+0 ratchet because 0x0096 was already counted by Wave 57's coverage of the
case-9 arm (StarbaseAcceptJobAction9_ReceivesBareJobAcceptReply). Wave 67
is depth coverage of a different case-arm with a load-bearing byte-exact
wire-shape assertion that case-9 does not have. Coverage stays at
102/207 (49.3%).

Touches:
  - server/src/PlayerConnection.cpp:10009-10013 (long -> int32_t job_id
    + 3-line explanatory comment)
  - tests/integration/CliClient.IntegrationTests/Opcodes/SectorStarbaseAcceptJob8Tests.cs (NEW)
  - tests/integration/CliClient.IntegrationTests/TestAccounts.cs
    (StarbaseAcceptJobAction8_OnSentinelJobId_ReceivesByteExact4ByteJobAcceptReply -> cli_test64)
  - tests/integration/CliClient.IntegrationTests/Fixtures/seed.sql
    (cli_test64 INSERT row, status=100)

Test runs in ~15s on the warmed-up live stack; build green; zero warnings.

+++ server/src/PlayerConnection.cpp
@@ -10006,7 +10006,10 @@ void Player::HandleStarbaseRequest(unsigned char *data)
 	case 8: // Accept job?
 		LogMessage("Accepting Job %d\n", pkt->StarbaseID);
 		{
-			long job_id = pkt->StarbaseID;
+			// Wire job_id field is canonical 4-byte int32; storing it as
+			// `long` would emit 8 bytes on LP64 platforms via the
+			// sizeof(job_id) below, diverging from the retail wire shape.
+			int32_t job_id = pkt->StarbaseID;
 			SendOpcode(ENB_OPCODE_0096_JOB_ACCEPT_REPLY, (u8*)&job_id, sizeof(job_id));
 			SectorManager *sm = GetSectorManager();
 			if (sm && !sm->AwardJob(this, job_id))
